requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
requests>=2.31.0
